requirements.txt
kafka-python==2.3.0
numpy==2.4.4
pandas==3.0.2
pyarrow==23.0.1
python-dotenv==1.2.2
duckdb==1.5.1
hiredis==3.3.1
redis==7.4.0
prometheus_client==0.24.1
